Personal-AI-Assistant/requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf
SpeechRecognition
pyaudio
pyttsx3